Introduce test mode for pre-deploy to avoid unnecessary operations on git pre-commit

+++ packages/site/pre-deploy.js
@@ -4,7 +4,7 @@ import {execSync} from 'child_process';
 
 import {createExec, logGreen, getEnv} from '../../tools/utils.js';
 
-const WORKER_ENV = getEnv().WORKER_ENV;
+const {WORKER_ENV, DO_INSTALL, DO_BUILD} = getEnv();
 
 const cwd = path.dirname(url.fileURLToPath(import.meta.url));;
 const exec = createExec(cwd);
@@ -13,7 +13,7 @@ const exec = createExec(cwd);
 logGreen('\n📦 {Site} Pre-deploy\n');
 
 // Install dependencies.
-exec('npm ci');
+DO_INSTALL && exec('npm ci');
 
 // Do TypeScript type-checking.
 exec('npx tsc --noEmit');
@@ -22,8 +22,8 @@ exec('npx tsc --noEmit');
 exec('npm test run');
 
 // Build.
-exec(`rm -rf dist`);
-exec(`npm run build`);
+DO_BUILD && exec(`rm -rf dist`);
+DO_BUILD && exec(`npm run build`);
 
 // Done.
 logGreen('\n📦 {Site} Pre-deploy script completed.\n');
